remove tile view component from source code
