+++ frontend/src/app/features/boards/openproject-boards.module.ts
@@ -53,7 +53,6 @@ import { SubprojectBoardHeaderComponent } from 'core-app/features/boards/board/b
 import { SubtasksBoardHeaderComponent } from 'core-app/features/boards/board/board-actions/subtasks/subtasks-board-header.component';
 import { StatusBoardHeaderComponent } from 'core-app/features/boards/board/board-actions/status/status-board-header.component';
 import { OpenprojectAutocompleterModule } from 'core-app/shared/components/autocompleter/openproject-autocompleter.module';
-import { TileViewComponent } from './tile-view/tile-view.component';
 
 @NgModule({
   imports: [
@@ -91,7 +90,6 @@ import { TileViewComponent } from './tile-view/tile-view.component';
     SubprojectBoardHeaderComponent,
     SubtasksBoardHeaderComponent,
     StatusBoardHeaderComponent,
-    TileViewComponent,
   ],
 })
 export class OpenprojectBoardsModule {
