--- frontend/src/app/features/boards/tile-view/tile-view.component.sass
@@ -1,44 +1,0 @@
-.op-tile-block
-  $block: &
-  display: grid
-  grid-template-rows: repeat(minmax(200px, auto))
-  grid-template-columns: auto auto
-  grid-gap: 1rem
-
-  &--tile
-    border-radius: 10px
-    display: grid
-    grid-template: 95px 1fr / 1fr
-    grid-template-columns: auto 1fr auto
-    grid-template-rows: auto
-    grid-row-gap: 5px
-    justify-items: left
-    background: #f7fafc
-    min-height: 150px
-
-    &:disabled
-      background: #fafafa
-
-    &:hover
-      text-decoration: none
-      border: 1px solid grey
-      border-radius: 10px !important
-      cursor: pointer
-
-  &--image
-    display: block
-    margin-top: auto
-    margin-bottom: auto
-
-  &--title
-    padding-top: 30px
-    padding-bottom: 5px
-    color: var(--primary-color-dark)
-    display: block
-    text-align: left
-    font-weight: var(--base-text-weight-bold)
-    font-size: large
-
-  &--description
-    text-align: left
-    width: 90%
